refactor: decompose SchedulingServiceImpl into single-responsibility services
