+++ src/main/kotlin/com/nickdferrara/fitify/scheduling/internal/listener/BusinessRuleEventListener.kt
@@ -1,6 +1,6 @@
 package com.nickdferrara.fitify.scheduling.internal.listener
 
-import com.nickdferrara.fitify.scheduling.internal.service.interfaces.SchedulingService
+import com.nickdferrara.fitify.scheduling.internal.service.interfaces.SchedulingCommandService
 import com.nickdferrara.fitify.shared.BusinessRuleUpdatedEvent
 import org.slf4j.LoggerFactory
 import org.springframework.context.event.EventListener
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component
 
 @Component
 internal class BusinessRuleEventListener(
-    private val schedulingService: SchedulingService,
+    private val schedulingCommandService: SchedulingCommandService,
 ) {
 
     private val logger = LoggerFactory.getLogger(BusinessRuleEventListener::class.java)
@@ -17,15 +17,15 @@ internal class BusinessRuleEventListener(
     fun onBusinessRuleUpdated(event: BusinessRuleUpdatedEvent) {
         when (event.ruleKey) {
             "cancellation_window_hours" -> {
-                schedulingService.cancellationWindowHours = event.newValue.toLong()
+                schedulingCommandService.cancellationWindowHours = event.newValue.toLong()
                 logger.info("Updated cancellationWindowHours to {}", event.newValue)
             }
             "max_waitlist_size" -> {
-                schedulingService.maxWaitlistSize = event.newValue.toInt()
+                schedulingCommandService.maxWaitlistSize = event.newValue.toInt()
                 logger.info("Updated maxWaitlistSize to {}", event.newValue)
             }
             "max_bookings_per_user_per_day" -> {
-                schedulingService.maxBookingsPerDay = event.newValue.toInt()
+                schedulingCommandService.maxBookingsPerDay = event.newValue.toInt()
                 logger.info("Updated maxBookingsPerDay to {}", event.newValue)
             }
         }
